Added 6 stomp-buttons and 6 adressable WS2811 RGB leds associated with each stomp-button.
The two right-most buttons are used to switch octave up/down. LEDs are colored to indicate the chosen octave. Octave 0 to 8 is supported, which is (only) almost as much as the MIDI standard allows, but should be more than enough for what most tone generators can support.
If a key is pressed while switching octave, the currently played note will persist until the key is released, after which pressing the key again will result in a new note in the new octave.

The two center buttons are used to switch program up/down, from program 0 to 12 (sent as program changes). LEDs are colored to indicate the chosen program.